pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"


[project]
name = "critiq"
version = "0.1.0"
dependencies = [
  "openai","tqdm","prettytable"
]
requires-python = ">= 3.10"
authors = [
  {name = "KYLN24", email = "1296845690@qq.com"},
]

[project.optional-dependencies]
vllm = ["vllm"]
sglang = ["sglang[all]"] # Manually specify --find-links https://flashinfer.ai/whl/cu124/torch2.4/flashinfer when `pip install` 
train = ["datasets", "trl>=0.12.0", "flash_attn", "deepspeed", "tensorboardX", "evaluate", "accelerate"]
dev = ["torch-tb-profiler", "ipykernel", "ipywidgets", ]
